server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
python-dotenv==1.0.1
openai==1.51.0
google-generativeai==0.8.0
PyPDF2==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.9.0
